refactor: MainMemo component 파일 분리
