+++ src/components/MainMemo/style.ts
@@ -3,8 +3,8 @@ import styled from 'styled-components';
 import * as T from './type';
 
 export const Div = styled.div<T.MemoTitle>`
-  width: 400px;
-  height: 400px;
+  width: ${(props) => props.width};
+  height: ${(props) => props.height};
   font-size: 20px;
   font-weight: bolder;
 `;
