+++ src/components/MainMemo/type.ts
@@ -1,3 +1,5 @@
 export type MemoTitle = {
+  width: string;
+  height: string;
   text: string;
 };
