+++ src/components/MemoDiv/style.ts
